Calculate Conditional Value at Risk (CVaR) for a set of returns.

This function computes CVaR using either quantiles or a normal distribution assumption. It can handle missing values and supports different confidence levels and distribution assumptions.

CVaR(returns, alpha = 0.05, dist = "quantile")

Arguments

returns

Numeric matrix or data frame of returns. Rows represent observations, and columns represent different assets or portfolios.

alpha

Confidence level for CVaR, typically between 0 and 1. Default is 0.05.

dist

Distribution assumption for CVaR calculation. Options are "quantile" (default) for quantile-based CVaR or "normal" for normal distribution assumption.

Value

A numeric vector or matrix of CVaR values corresponding to each column in the input returns.

Examples

if (FALSE) {
library(xts)
# Assuming xts objects
returns_xts <- xts(matrix(c(0.01, 0.03, -0.01, 0.04, -0.06, -0.02, 0.03,
 -0.01, 0.02, -0.03, 0.03, -0.006, 0.02, -0.03), ncol = 2),
 order.by = as.Date(1:7))

# Calculate quantile-based CVaR
CVaR(returns_xts, alpha = 0.05, dist = "quantile")

# Calculate CVaR assuming normal distribution
CVaR(returns_xts, alpha = 0.05, dist = "normal")
}